snackhub2/pyproject.toml
[project]
name = "snackhub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "flet>=0.28.3",
    "mysql-connector-python>=9.5.0",
]
